Move worker info to WorkerManager
